Update MSVC toolchain to 19.51.36244 and win_clang to 23.0
$WIN_TOOLCHAIN/win_sdk is now $WIN_TOOLCHAIN/Windows Kits/10
so folks will have to update their GN args (like we had to do
for these tests).

The FASTLINK option was removed from MSVC 2026, so I removed it here
https://github.com/bazelbuild/rules_cc/issues/558

I changed the create.py script to be runnable on Linux because
I didn't have my Windows box set up to upload things via gcloud
and the current iteration of the script does not need to run on
Windows (although the depot_tools script probably does).

I left some breadcrumbs with the Bazel toolchain (Windows+Bazel
is not working atm) for when I get around to fixing that.

The latest MSVC headers need Clang 20+ to work, so this updates
us to use the latest version Chromium has built (version 23 TOT).
This new compiler brought some new warnings, some of which I
squashed and others which I fixed.

Finally, I added MSVC jobs back to the CQ (they were temporarily
removed when broken).

Bug: 514697001
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1245496
Commit-Queue: Jorge Betancourt <jmbetancourt@google.com>
Auto-Submit: Kaylee Lubick <kjlubick@google.com>

+++ docs/examples/Matrix_getType.cpp
@@ -5,8 +5,8 @@ REG_FIDDLE(Matrix_getType, 256, 256, true, 0) {
 void draw(SkCanvas* canvas) {
     SkMatrix matrix;
     matrix.setAll(1, 0, 0,   0, 1, 0,    0, 0, 1);
-    SkDebugf("identity flags hex: %0x decimal: %d\n", matrix.getType(), matrix.getType());
+    SkDebugf("identity flags hex: %0x decimal: %d\n", (unsigned)matrix.getType(), matrix.getType());
     matrix.setAll(1, 0, 0,   0, 1, 0,    0, 0, .5f);
-    SkDebugf("set all  flags hex: %0x decimal: %d\n", matrix.getType(), matrix.getType());
+    SkDebugf("set all  flags hex: %0x decimal: %d\n", (unsigned)matrix.getType(), matrix.getType());
 }
 }  // END FIDDLE
